pyproject.toml
[tool.poetry]
name = "pigeongpt"
version = "0.1.0"
description = ""
authors = ["Kal.L <kalvin807@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
google-api-python-client = "^2.85.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^1.0.0"
openai = "^0.27.4"
python-dotenv = "^1.0.0"
tiktoken = "^0.3.3"
tenacity = "^8.2.2"
langchain = "^0.0.141"
bs4 = "^0.0.1"
lxml = "^4.9.2"
structlog = "^23.1.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
ruff = "^0.0.261"
pytest = "^7.3.0"
mypy = "^1.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
